created a SelectItem common component and refactored
