+++ src/components/common/SelectItem.tsx
@@ -1,0 +1,35 @@
+import { CheckIcon } from "@radix-ui/react-icons";
+import * as Select from "@radix-ui/react-select";
+import classnames from "classnames";
+import React, { type ReactNode } from "react";
+
+type SelectItemProps = {
+  children: ReactNode;
+  className?: string;
+  value: string;
+};
+
+const SelectItem = React.forwardRef<HTMLDivElement, SelectItemProps>(
+  ({ children, className, value, ...props }, forwardedRef) => {
+    return (
+      <Select.Item
+        className={classnames(
+          "relative flex h-6 items-center justify-center rounded px-7 py-0 text-base data-[highlighted]:bg-blue-600 data-[highlighted]:text-white data-[highlighted]:outline-none",
+          className
+        )}
+        ref={forwardedRef}
+        value={value}
+        {...props}
+      >
+        <Select.ItemText>{children}</Select.ItemText>
+        <Select.ItemIndicator className="absolute left-0 inline-flex w-6 items-center justify-center">
+          <CheckIcon />
+        </Select.ItemIndicator>
+      </Select.Item>
+    );
+  }
+);
+
+SelectItem.displayName = "SelectItem";
+
+export default SelectItem;
