+++ src/components/siteDiary/weather/Dropdown.tsx
@@ -1,47 +1,15 @@
 import type { WeatherCondition } from "@prisma/client";
-import {
-  CheckIcon,
-  ChevronDownIcon,
-  ChevronUpIcon,
-} from "@radix-ui/react-icons";
+import { ChevronDownIcon, ChevronUpIcon } from "@radix-ui/react-icons";
 import * as Select from "@radix-ui/react-select";
-import classnames from "classnames";
-import React, { type ReactNode } from "react";
+import dynamic from "next/dynamic";
 import { Rainy } from "styled-icons/ionicons-outline";
-
-type SelectItemProps = {
-  children: ReactNode;
-  className?: string;
-  value: string;
-};
 
 type DropdownProps = {
   weatherCondition: WeatherCondition | null | undefined;
   onWeatherChange: (value: WeatherCondition) => void;
 };
 
-const SelectItem = React.forwardRef<HTMLDivElement, SelectItemProps>(
-  ({ children, className, value, ...props }, forwardedRef) => {
-    return (
-      <Select.Item
-        className={classnames(
-          "relative flex h-6 items-center justify-center rounded px-7 py-0 text-base data-[highlighted]:bg-blue-600 data-[highlighted]:text-white data-[highlighted]:outline-none",
-          className
-        )}
-        ref={forwardedRef}
-        value={value}
-        {...props}
-      >
-        <Select.ItemText>{children}</Select.ItemText>
-        <Select.ItemIndicator className="absolute left-0 inline-flex  items-center justify-center">
-          <CheckIcon />
-        </Select.ItemIndicator>
-      </Select.Item>
-    );
-  }
-);
-
-SelectItem.displayName = "SelectItem";
+const SelectItem = dynamic(() => import("../../common/SelectItem"));
 
 const Dropdown = ({ weatherCondition, onWeatherChange }: DropdownProps) => {
   return (
